Move Settings component to shared-components for better modularity.
