+++ playgrounds/next/src/app/layout.tsx
@@ -1,9 +1,8 @@
 import { Navigation } from '@/components/Navigation';
-import { Settings } from '@/components/Settings';
 import Providers from '@/lib/providers';
+import { Settings } from 'shared-components';
 import { ssrClient } from './marketplace-sdk/ssr';
 import '@0xsequence/design-system/index.css';
-import '@0xsequence/kit/styles.css';
 import './globals.css';
 
 export default async function RootLayout({
