+++ playgrounds/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from './store';
 export * from './consts';
 export * from './types';
+export * from './components';
